型のフォルダを作成
